+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ## 2.0.0 under development
 
 - Chg #63: Raise PHP version to 8.0 (@arogachev)
-- Enh #63: Improve perfomance (@arogachev)
+- Enh #63: Improve performance (@arogachev)
+- Enh #?: Implement `getByNames()` and `getAccessTree()` methods in `ItemsStorage` (@arogachev)
+- Enh #?: Implement `filterUserItemNames()` method in `AssignmentsStorage` (@arogachev)
+- Chg #?: Rename `$name` argument to `$names` and allow array type for it in `getAllChildren()`, `getAllChildRoles()`,
+  `getAllChildPermissions()` methods in `ItemsStorage` (@arogachev)
 
 ## 1.0.0 April 08, 2022